requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
contextlib2==0.5.5
cycler==0.10.0
docopt==0.6.2
docx2txt==0.6
EbookLib==0.17.1
entrypoints==0.3
flake8==3.7.7
Flask==1.0.3
fulltext==0.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
lxml==4.3.3
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mock==3.0.5
mpmath==1.1.0
numpy==1.16.3
Pillow==6.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.0
pytesseract==0.2.6
python-dateutil==2.8.0
python-docx==0.8.10
python-Levenshtein==0.12.0
python-magic==0.4.15
python-pptx==0.6.5
rarfile==3.0
requests==2.22.0
scipy==1.2.1
six==1.10.0
SpeechRecognition==3.6.3
sympy==1.4
tika==1.19
urllib3==1.25.2
Werkzeug==0.15.4
xlrd==1.0.0
XlsxWriter==1.1.8
